Creating SchoolInfoScreen style file
